# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisConnect API — hospitality emergency response coordination
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EmergencyType = (typeof EmergencyType)[keyof typeof EmergencyType];

export const EmergencyType = {
  fire: "fire",
  medical: "medical",
  security: "security",
  other: "other",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export type AlertReportedBy =
  (typeof AlertReportedBy)[keyof typeof AlertReportedBy];

export const AlertReportedBy = {
  guest: "guest",
  staff: "staff",
  admin: "admin",
} as const;

export interface Alert {
  id: string;
  name: string;
  location: string;
  type: EmergencyType;
  description: string;
  status: AlertStatus;
  timestamp: string;
  reportedBy: AlertReportedBy;
}

export type CreateAlertInputReportedBy =
  (typeof CreateAlertInputReportedBy)[keyof typeof CreateAlertInputReportedBy];

export const CreateAlertInputReportedBy = {
  guest: "guest",
  staff: "staff",
  admin: "admin",
} as const;

export interface CreateAlertInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  location: string;
  type: EmergencyType;
  description: string;
  reportedBy: CreateAlertInputReportedBy;
}

export interface UpdateAlertInput {
  status: AlertStatus;
}

export type AlertStatsByType = {
  fire: number;
  medical: number;
  security: number;
  other: number;
};

export interface AlertStats {
  total: number;
  active: number;
  pending: number;
  inProgress: number;
  resolved: number;
  byType: AlertStatsByType;
}

export type ListAlertsParams = {
  type?: EmergencyType;
  status?: AlertStatus;
};