# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CrisisConnect API — hospitality emergency response coordination
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: alerts
    description: Emergency alert operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all alerts
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EmergencyType"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AlertStatus"
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a new emergency alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    put:
      operationId: updateAlertStatus
      tags: [alerts]
      summary: Update an alert's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/stats:
    get:
      operationId: getAlertStats
      tags: [alerts]
      summary: Get aggregated alert statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertStats"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    EmergencyType:
      type: string
      enum: [fire, medical, security, other]
    AlertStatus:
      type: string
      enum: [pending, in_progress, resolved]
    Alert:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        type:
          $ref: "#/components/schemas/EmergencyType"
        description:
          type: string
        status:
          $ref: "#/components/schemas/AlertStatus"
        timestamp:
          type: string
          format: date-time
        reportedBy:
          type: string
          enum: [guest, staff, admin]
      required: [id, name, location, type, description, status, timestamp, reportedBy]
    CreateAlertInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        type:
          $ref: "#/components/schemas/EmergencyType"
        description:
          type: string
        reportedBy:
          type: string
          enum: [guest, staff, admin]
      required: [name, location, type, description, reportedBy]
    UpdateAlertInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/AlertStatus"
      required: [status]
    AlertStats:
      type: object
      properties:
        total:
          type: integer
        active:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        resolved:
          type: integer
        byType:
          type: object
          properties:
            fire:
              type: integer
            medical:
              type: integer
            security:
              type: integer
            other:
              type: integer
          required: [fire, medical, security, other]
      required: [total, active, pending, inProgress, resolved, byType]