# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisConnect API — hospitality emergency response coordination
 * OpenAPI spec version: 0.1.0
 */
import type { AlertReportedBy } from "./alertReportedBy";
import type { AlertStatus } from "./alertStatus";
import type { EmergencyType } from "./emergencyType";

export interface Alert {
  id: string;
  name: string;
  location: string;
  type: EmergencyType;
  description: string;
  status: AlertStatus;
  timestamp: Date;
  reportedBy: AlertReportedBy;
}