# lib/api-zod/src/generated/types/alertStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisConnect API — hospitality emergency response coordination
 * OpenAPI spec version: 0.1.0
 */
import type { AlertStatsByType } from "./alertStatsByType";

export interface AlertStats {
  total: number;
  active: number;
  pending: number;
  inProgress: number;
  resolved: number;
  byType: AlertStatsByType;
}