# lib/api-zod/src/generated/types/createAlertInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisConnect API — hospitality emergency response coordination
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAlertInputReportedBy } from "./createAlertInputReportedBy";
import type { EmergencyType } from "./emergencyType";

export interface CreateAlertInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  location: string;
  type: EmergencyType;
  description: string;
  reportedBy: CreateAlertInputReportedBy;
}