# lib/api-zod/src/generated/types/emergencyType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisConnect API — hospitality emergency response coordination
 * OpenAPI spec version: 0.1.0
 */

export type EmergencyType = (typeof EmergencyType)[keyof typeof EmergencyType];

export const EmergencyType = {
  fire: "fire",
  medical: "medical",
  security: "security",
  other: "other",
} as const;